cleaned up error handling for dropped messages
right now, the user is notified that messages were dropped, but no other
action is taken. It's assumed that the user is at least storing trials
to disk on the beaglebone or is not doing so knowingly. The main thing
we want to avoid is having an experiment crash because the host did, as
that would put the animal at risk.